:zap: Implement more efficient maximal filtering
Generate maximal subset of the powerset of extensions by maintaining a partial set of maximal entensions and considering each extension only once, removing those proven to not be maximal and adding the considered extension if it itself proved to be maximal wrt the current parial-set.

Adjust the SE-PR task method accordingly.